requirements.txt
Django
Markdown
PyYAML
South
argparse
defusedxml
dj-database-url
dj-static
django-debug-toolbar
django-dotenv
django-extensions
django-filter
django-getenv
django-oauth-plus
django-oauth2-provider
django-storages
django-toolbelt
djangorestframework
gunicorn
httplib2
ipython
newrelic
oauth2
oauthlib
pillow
psycopg2
python-linkedin
raven
requests
requests-oauthlib
shortuuid
six
static
wsgiref
wiki